backend/requirements.txt
djangorestframework
markdown
drf-yasg
django-filter
django-tinymce
Pygments
